Move Registrate builders to `com.gregtechceu.gtceu.api.registry.registrate.builder`
