+++ docs/content/Modpacks/Other-Topics/Ambiguous-Methods.md
@@ -18,9 +18,9 @@ StartupEvents.registry('gtceu:machine', event => {
 
 you'd get the error:
 ```
-Error in 'StartupEvents.registry': The choice of Java method com.gregtechceu.gtceu.api.registry.registrate.MultiblockMachineBuilder.tooltips matching JavaScript argument types (net.minecraft.network.chat.MutableComponent) is ambiguous; candidate methods are: 
-    class com.gregtechceu.gtceu.api.registry.registrate.MachineBuilder tooltips(java.util.List)
-    class com.gregtechceu.gtceu.api.registry.registrate.MachineBuilder tooltips(net.minecraft.network.chat.Component[])
+Error in 'StartupEvents.registry': The choice of Java method com.gregtechceu.gtceu.api.registry.registrate.builder.MultiblockMachineBuilder.tooltips matching JavaScript argument types (net.minecraft.network.chat.MutableComponent) is ambiguous; candidate methods are: 
+    class com.gregtechceu.gtceu.api.registry.registrate.builder.MachineBuilder tooltips(java.util.List)
+    class com.gregtechceu.gtceu.api.registry.registrate.builder.MachineBuilder tooltips(net.minecraft.network.chat.Component[])
 ```
 
 In this case, there's ambiguity between the following 2 java functions:
