+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/GTRegistrate.java
@@ -22,7 +22,8 @@ import com.gregtechceu.gtceu.api.machine.multiblock.MultiblockControllerMachine;
 import com.gregtechceu.gtceu.api.recipe.GTRecipeType;
 import com.gregtechceu.gtceu.api.recipe.category.GTRecipeCategory;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
-import com.gregtechceu.gtceu.api.registry.registrate.forge.GTFluidBuilder;
+import com.gregtechceu.gtceu.api.registry.registrate.builder.*;
+import com.gregtechceu.gtceu.api.registry.registrate.builder.forge.GTFluidBuilder;
 import com.gregtechceu.gtceu.api.registry.registrate.holder.HolderRegistryEntry;
 import com.gregtechceu.gtceu.api.registry.registrate.holder.NoConfigHolderBuilder;
 import com.gregtechceu.gtceu.core.mixins.registrate.AbstractRegistrateAccessor;
