+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/builder/IGTFluidBuilder.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.api.registry.registrate;
+package com.gregtechceu.gtceu.api.registry.registrate.builder;
 
 import com.gregtechceu.gtceu.api.fluids.FluidState;
