+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/builder/forge/GTFluidBuilder.java
@@ -1,9 +1,10 @@
-package com.gregtechceu.gtceu.api.registry.registrate.forge;
+package com.gregtechceu.gtceu.api.registry.registrate.builder.forge;
 
 import com.gregtechceu.gtceu.api.data.chemical.material.Material;
 import com.gregtechceu.gtceu.api.fluids.FluidState;
 import com.gregtechceu.gtceu.api.fluids.GTFluid;
-import com.gregtechceu.gtceu.api.registry.registrate.IGTFluidBuilder;
+import com.gregtechceu.gtceu.api.registry.registrate.GTClientFluidTypeExtensions;
+import com.gregtechceu.gtceu.api.registry.registrate.builder.IGTFluidBuilder;
 import com.gregtechceu.gtceu.common.item.GTBucketItem;
 import com.gregtechceu.gtceu.utils.GTUtil;
